+++ agentwire/tts/engines/__init__.py
@@ -1,6 +1,7 @@
 """TTS Engine Implementations"""
 
 from .chatterbox import ChatterboxEngine
+from .kokoro import KokoroEngine
 from .qwen_base import QwenBaseEngine
 from .qwen_custom import QwenCustomEngine
 from .qwen_design import QwenDesignEngine
@@ -8,6 +9,7 @@ from .zonos import ZonosHybridEngine, ZonosTransformerEngine
 
 __all__ = [
     "ChatterboxEngine",
+    "KokoroEngine",
     "QwenBaseEngine",
     "QwenCustomEngine",
     "QwenDesignEngine",